Refactoring and cleanup (removing unecessary packages). Fixed a bug in extension flattening.
