+++ encoders/src/test/java/au/csiro/pathling/encoders/SchemaConverterTest.java
@@ -11,12 +11,11 @@
  *
  */
 
-package au.csiro.pathling.encoders2;
+package au.csiro.pathling.encoders;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import au.csiro.pathling.encoders.EncoderConfig;
 import au.csiro.pathling.encoders.datatypes.DataTypeMappings;
 import au.csiro.pathling.encoders.datatypes.R4DataTypeMappings;
 import ca.uhn.fhir.context.FhirContext;
