+++ fhir-server/src/main/java/au/csiro/pathling/fhirpath/element/ExtensionPath.java
@@ -6,7 +6,7 @@
 
 package au.csiro.pathling.fhirpath.element;
 
-import au.csiro.pathling.encoders2.ExtensionSupport;
+import au.csiro.pathling.encoders.ExtensionSupport;
 import au.csiro.pathling.fhirpath.ResourcePath;
 import java.util.Arrays;
 import java.util.List;
