+++ fhir-server/src/main/java/au/csiro/pathling/fhirpath/function/ExtensionFunction.java
@@ -8,7 +8,7 @@ package au.csiro.pathling.fhirpath.function;
 
 import static au.csiro.pathling.utilities.Preconditions.checkUserInput;
 
-import au.csiro.pathling.encoders2.ExtensionSupport;
+import au.csiro.pathling.encoders.ExtensionSupport;
 import au.csiro.pathling.fhirpath.Comparable.ComparisonOperation;
 import au.csiro.pathling.fhirpath.FhirPath;
 import au.csiro.pathling.fhirpath.NonLiteralPath;
